+++ src/components/FeatureStore/FeatureStore.js
@@ -1,23 +1,24 @@
 import React, { useCallback, useEffect, useState, useRef } from 'react'
 import { connect } from 'react-redux'
 import PropTypes from 'prop-types'
-import axios from 'axios'
 
 import Loader from '../../common/Loader/Loader'
 import Content from '../../layout/Content/Content'
 import RegisterArtifactPopup from '../RegisterArtifactPopup/RegisterArtifactPopup'
 
 import artifactsAction from '../../actions/artifacts'
-import { generateArtifacts } from '../../utils/generateArtifacts'
-import { detailsMenu, generatePageData } from './feaureStore.util'
+import {
+  checkTabIsValid,
+  detailsMenu,
+  generatePageData,
+  handleApplyDetailsChanges,
+  handleFetchData,
+  navigateToDetailsPane
+} from './feaureStore.util'
 import { handleArtifactTreeFilterChange } from '../../utils/handleArtifactTreeFilterChange'
 import {
   DATASETS_TAB,
   FEATURE_SETS_TAB,
-  FEATURES_TAB,
-  DETAILS_ANALYSIS_TAB,
-  DETAILS_METADATA_TAB,
-  DETAILS_STATISTICS_TAB,
   FEATURE_VECTORS_TAB
 } from '../../constants'
 import notificationActions from '../../actions/notification'
@@ -56,50 +57,22 @@ const FeatureStore = ({
 
   const fetchData = useCallback(
     async item => {
-      let result
+      const data = await handleFetchData(
+        featureStoreRef,
+        fetchDataSets,
+        fetchFeatureSets,
+        fetchFeatures,
+        fetchFeatureVectors,
+        item,
+        match.params.pageTab
+      )
 
-      if (match.params.pageTab === DATASETS_TAB) {
-        result = await fetchDataSets(item)
-
-        if (result) {
-          setContent(generateArtifacts(result))
-          setYamlContent(result)
-
-          return result
-        }
-      } else if (match.params.pageTab === FEATURE_SETS_TAB) {
-        const config = {
-          cancelToken: new axios.CancelToken(cancel => {
-            featureStoreRef.current.cancel = cancel
-          })
-        }
-
-        result = await fetchFeatureSets(item, config)
-
-        if (result) {
-          setContent(result)
-          setYamlContent(result)
-
-          return result
-        }
-      } else if (match.params.pageTab === FEATURES_TAB) {
-        result = await fetchFeatures(item)
-
-        if (result) {
-          setContent(result)
-          setYamlContent(result)
-
-          return result
-        }
-      } else if (match.params.pageTab === FEATURE_VECTORS_TAB) {
-        result = await fetchFeatureVectors(item)
-
-        if (result) {
-          setContent(result)
-
-          return result
-        }
+      if (data.content.length > 0) {
+        setContent(data.content)
+        setYamlContent(data.yamlContent)
       }
+
+      return data.yamlContent
     },
     [
       match.params.pageTab,
@@ -145,6 +118,7 @@ const FeatureStore = ({
               loading: false
             }
           }))
+          setYamlContent(result)
         })
         .catch(error =>
           setPageData(state => ({
@@ -158,6 +132,15 @@ const FeatureStore = ({
         )
     },
     [fetchFeatureVector, match.params.projectName]
+  )
+
+  const handleExpandRow = useCallback(
+    (item, isCollapse) => {
+      if (match.params.pageTab === FEATURE_VECTORS_TAB && isCollapse) {
+        setYamlContent(artifactsStore.featureVectors.allData)
+      }
+    },
+    [artifactsStore.featureVectors.allData, match.params.pageTab]
   )
 
   useEffect(() => {
@@ -204,73 +187,7 @@ const FeatureStore = ({
   }, [handleRemoveFeatureVector, handleRequestOnExpand, match.params.pageTab])
 
   useEffect(() => {
-    const { name, tag } = match.params
-    let artifacts = []
-
-    if (
-      match.params.pageTab === FEATURE_SETS_TAB &&
-      artifactsStore.featureSets.length > 0
-    ) {
-      artifacts = artifactsStore.featureSets
-    } else if (
-      match.params.pageTab === FEATURES_TAB &&
-      artifactsStore.features.length > 0
-    ) {
-      artifacts = artifactsStore.features
-    } else if (
-      match.params.pageTab === DATASETS_TAB &&
-      artifactsStore.dataSets.length > 0
-    ) {
-      artifacts = artifactsStore.dataSets
-    } else if (
-      match.params.pageTab === FEATURE_VECTORS_TAB &&
-      artifactsStore.featureVectors.allData.length > 0
-    ) {
-      if (artifactsStore.featureVectors.selectedRowData.content[name]) {
-        artifacts = artifactsStore.featureVectors.selectedRowData.content[name]
-      } else {
-        artifacts = artifactsStore.featureVectors.allData
-      }
-    }
-
-    if (match.params.name && artifacts.length !== 0) {
-      const [selectedArtifact] = artifacts.filter(artifact => {
-        const searchKey = artifact.name ? 'name' : 'key'
-
-        if (
-          match.params.pageTab === FEATURE_SETS_TAB ||
-          match.params.pageTab === FEATURE_VECTORS_TAB
-        ) {
-          return artifact[searchKey] === name && artifact.tag === tag
-        } else {
-          return artifact[searchKey] === name
-        }
-      })
-
-      if (!selectedArtifact) {
-        history.push(
-          `/projects/${match.params.projectName}/feature-store/${match.params.pageTab}`
-        )
-      } else {
-        if (match.params.pageTab === DATASETS_TAB) {
-          const [dataSet] = selectedArtifact.data.filter(item => {
-            if (selectedArtifact.link_iteration) {
-              const { link_iteration } = selectedArtifact.link_iteration
-
-              return link_iteration === item.iter
-            }
-
-            return true
-          })
-
-          return setSelectedItem({ item: dataSet })
-        }
-
-        setSelectedItem({ item: selectedArtifact })
-      }
-    } else {
-      setSelectedItem({})
-    }
+    navigateToDetailsPane(artifactsStore, history, match, setSelectedItem)
   }, [
     match.params,
     artifactsStore.artifacts,
@@ -278,28 +195,13 @@ const FeatureStore = ({
     artifactsStore.dataSets,
     artifactsStore.featureSets,
     artifactsStore.features,
-    artifactsStore.featureVectors
+    artifactsStore.featureVectors,
+    artifactsStore,
+    match
   ])
 
   useEffect(() => {
-    if (
-      (match.params.tab?.toUpperCase() === DETAILS_METADATA_TAB &&
-        !selectedItem.item?.schema &&
-        !selectedItem.item?.entities) ||
-      (match.params.tab === DETAILS_ANALYSIS_TAB &&
-        !selectedItem.item?.extra_data) ||
-      (match.params.tab?.toUpperCase() === DETAILS_STATISTICS_TAB &&
-        match.params.pageTab !== FEATURE_SETS_TAB &&
-        !selectedItem.item?.entities)
-    ) {
-      history.push(
-        `/projects/${match.params.projectName}/feature-store/${
-          match.params.pageTab
-        }/${match.params.name}${
-          match.params.tag ? `/${match.params.tag}` : ''
-        }/overview`
-      )
-    }
+    checkTabIsValid(history, match, selectedItem)
 
     setPageData(state => {
       const newDetailsMenu = [...detailsMenu]
@@ -344,7 +246,9 @@ const FeatureStore = ({
     selectedItem.item,
     match.params.pageTab,
     match.params.tag,
-    selectedItem.entities
+    selectedItem.entities,
+    match,
+    selectedItem
   ])
 
   const handleTreeFilterChange = useCallback(
@@ -375,49 +279,14 @@ const FeatureStore = ({
   )
 
   const applyDetailsChanges = changes => {
-    const data = {
-      spec: {
-        ...changes.data
-      }
-    }
-
-    if (data.spec.labels) {
-      const objectLabels = {}
-
-      data.spec.labels.forEach(label => {
-        const splitedLabel = label.split(':')
-
-        objectLabels[splitedLabel[0]] = splitedLabel[1].replace(' ', '')
-      })
-
-      data.spec.labels = { ...objectLabels }
-    }
-
-    return updateFeatureSetData(
-      match.params.projectName,
-      match.params.name,
-      selectedItem.item.tag,
-      data
+    handleApplyDetailsChanges(
+      changes,
+      fetchData,
+      match,
+      selectedItem,
+      setNotification,
+      updateFeatureSetData
     )
-      .then(response => {
-        return fetchData({ project: match.params.projectName }).then(() => {
-          setNotification({
-            status: response.status,
-            id: Math.random(),
-            message: 'Updated successfully'
-          })
-
-          return response
-        })
-      })
-      .catch(error => {
-        setNotification({
-          status: 400,
-          id: Math.random(),
-          message: 'Fail to update',
-          retry: applyDetailsChanges
-        })
-      })
   }
 
   return (
@@ -430,6 +299,7 @@ const FeatureStore = ({
             featureStoreRef.current.cancel(message)
         }}
         content={content}
+        expandRow={handleExpandRow}
         groupFilter={groupFilter}
         handleArtifactFilterTree={handleTreeFilterChange}
         handleCancel={() => setSelectedItem({})}
