+++ src/components/FeatureStore/feaureStore.util.js
@@ -1,8 +1,14 @@
 import {
+  DATASETS_TAB,
+  DETAILS_ANALYSIS_TAB,
+  DETAILS_METADATA_TAB,
+  DETAILS_STATISTICS_TAB,
   FEATURE_SETS_TAB,
   FEATURE_VECTORS_TAB,
   FEATURES_TAB
 } from '../../constants'
+import { generateArtifacts } from '../../utils/generateArtifacts'
+import axios from 'axios'
 
 export const datasetsInfoHeaders = [
   { label: 'Hash', id: 'hash' },
@@ -211,4 +217,209 @@ export const generatePageData = (
   }
 
   return data
+}
+
+export const handleFetchData = async (
+  featureStoreRef,
+  fetchDataSets,
+  fetchFeatureSets,
+  fetchFeatures,
+  fetchFeatureVectors,
+  item,
+  pageTab
+) => {
+  let data = {
+    content: [],
+    yamlContent: []
+  }
+  let result = null
+
+  if (pageTab === DATASETS_TAB) {
+    result = await fetchDataSets(item)
+
+    if (result) {
+      data.content = generateArtifacts(result)
+      data.yamlContent = result
+    }
+  } else if (pageTab === FEATURE_SETS_TAB) {
+    const config = {
+      cancelToken: new axios.CancelToken(cancel => {
+        featureStoreRef.current.cancel = cancel
+      })
+    }
+
+    result = await fetchFeatureSets(item, config)
+
+    if (result) {
+      data.content = result
+      data.yamlContent = result
+    }
+  } else if (pageTab === FEATURES_TAB) {
+    result = await fetchFeatures(item)
+
+    if (result) {
+      data.content = result
+      data.yamlContent = result
+    }
+  } else if (pageTab === FEATURE_VECTORS_TAB) {
+    result = await fetchFeatureVectors(item)
+    result[0].labels = { 1: '1' }
+
+    if (result) {
+      data.content = result
+      data.yamlContent = result
+    }
+  }
+
+  return data
+}
+
+export const navigateToDetailsPane = (
+  artifactsStore,
+  history,
+  match,
+  setSelectedItem
+) => {
+  const { name, tag } = match.params
+  let artifacts = []
+
+  if (
+    match.params.pageTab === FEATURE_SETS_TAB &&
+    artifactsStore.featureSets.length > 0
+  ) {
+    artifacts = artifactsStore.featureSets
+  } else if (
+    match.params.pageTab === FEATURES_TAB &&
+    artifactsStore.features.length > 0
+  ) {
+    artifacts = artifactsStore.features
+  } else if (
+    match.params.pageTab === DATASETS_TAB &&
+    artifactsStore.dataSets.length > 0
+  ) {
+    artifacts = artifactsStore.dataSets
+  } else if (
+    match.params.pageTab === FEATURE_VECTORS_TAB &&
+    artifactsStore.featureVectors.allData.length > 0
+  ) {
+    if (artifactsStore.featureVectors.selectedRowData.content[name]) {
+      artifacts = artifactsStore.featureVectors.selectedRowData.content[name]
+    } else {
+      artifacts = artifactsStore.featureVectors.allData
+    }
+  }
+
+  if (match.params.name && artifacts.length !== 0) {
+    const [selectedArtifact] = artifacts.filter(artifact => {
+      const searchKey = artifact.name ? 'name' : 'key'
+
+      if (
+        match.params.pageTab === FEATURE_SETS_TAB ||
+        match.params.pageTab === FEATURE_VECTORS_TAB
+      ) {
+        return artifact[searchKey] === name && artifact.tag === tag
+      } else {
+        return artifact[searchKey] === name
+      }
+    })
+
+    if (!selectedArtifact) {
+      history.push(
+        `/projects/${match.params.projectName}/feature-store/${match.params.pageTab}`
+      )
+    } else {
+      if (match.params.pageTab === DATASETS_TAB) {
+        const [dataSet] = selectedArtifact.data.filter(item => {
+          if (selectedArtifact.link_iteration) {
+            const { link_iteration } = selectedArtifact.link_iteration
+
+            return link_iteration === item.iter
+          }
+
+          return true
+        })
+
+        return setSelectedItem({ item: dataSet })
+      }
+
+      setSelectedItem({ item: selectedArtifact })
+    }
+  } else {
+    setSelectedItem({})
+  }
+}
+
+export const handleApplyDetailsChanges = (
+  changes,
+  fetchData,
+  match,
+  selectedItem,
+  setNotification,
+  updateFeatureSetData
+) => {
+  const data = {
+    spec: {
+      ...changes.data
+    }
+  }
+
+  if (data.spec.labels) {
+    const objectLabels = {}
+
+    data.spec.labels.forEach(label => {
+      const splitedLabel = label.split(':')
+
+      objectLabels[splitedLabel[0]] = splitedLabel[1].replace(' ', '')
+    })
+
+    data.spec.labels = { ...objectLabels }
+  }
+
+  return updateFeatureSetData(
+    match.params.projectName,
+    match.params.name,
+    selectedItem.item.tag,
+    data
+  )
+    .then(response => {
+      return fetchData({ project: match.params.projectName }).then(() => {
+        setNotification({
+          status: response.status,
+          id: Math.random(),
+          message: 'Updated successfully'
+        })
+
+        return response
+      })
+    })
+    .catch(error => {
+      setNotification({
+        status: 400,
+        id: Math.random(),
+        message: 'Fail to update',
+        retry: handleApplyDetailsChanges
+      })
+    })
+}
+
+export const checkTabIsValid = (history, match, selectedItem) => {
+  if (
+    (match.params.tab?.toUpperCase() === DETAILS_METADATA_TAB &&
+      !selectedItem.item?.schema &&
+      !selectedItem.item?.entities) ||
+    (match.params.tab?.toUpperCase() === DETAILS_ANALYSIS_TAB &&
+      match.params.pageTab !== FEATURE_SETS_TAB &&
+      !selectedItem.item?.extra_data) ||
+    (match.params.tab?.toUpperCase() === DETAILS_STATISTICS_TAB &&
+      match.params.pageTab !== FEATURE_SETS_TAB &&
+      !selectedItem.item?.entities)
+  ) {
+    return history.push(
+      `/projects/${match.params.projectName}/feature-store/${
+        match.params.pageTab
+      }/${match.params.name}${
+        match.params.tag ? `/${match.params.tag}` : ''
+      }/overview`
+    )
+  }
 }
